requirements.txt
panda3d==1.10.15
pillow